pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_workshop"
version = "0.1.0"
description = "AI Workshop API Service"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "httpx>=0.24.0",
    "python-dotenv>=0.19.0",
    "fastmcp>=0.1.0"
]
